Main Activity buttons change, min/max resize of map

+++ app/src/main/java/pl/edu/ug/saveplaceonmap/models/Location.java
@@ -3,13 +3,13 @@ package pl.edu.ug.saveplaceonmap.models;
 public class Location {
 
     private int id;
-    private float x;
-    private float y;
+    private double x;
+    private double y;
     private String title;
     private String description;
     private Category category;
 
-    public Location(int id, float x, float y, String title, String description, Category category) {
+    public Location(int id, double x, double y, String title, String description, Category category) {
         this.id = id;
         this.x = x;
         this.y = y;
@@ -22,11 +22,11 @@ public class Location {
         return id;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
